requirements.txt
django
django-admin-interface
django-smart-selects